+++ app/models/user_item.rb
@@ -1,4 +1,6 @@
 class UserItem < ApplicationRecord
   belongs_to :user
   belongs_to :item
+
+  enum quality: { normal: 'normal', rare: 'rare', epic: 'epic', legendary: 'legendary' }
 end