+++ db/migrate/20250906153638_add_experimental_crafting_to_users.rb
@@ -1,0 +1,5 @@
+class AddExperimentalCraftingToUsers < ActiveRecord::Migration[8.0]
+  def change
+    add_column :users, :experimental_crafting, :boolean, default: false, null: false
+  end
+end