+++ db/migrate/20250906153642_add_quality_to_user_items.rb
@@ -1,0 +1,5 @@
+class AddQualityToUserItems < ActiveRecord::Migration[8.0]
+  def change
+    add_column :user_items, :quality, :string, default: 'normal', null: false
+  end
+end